+++ init.lua
@@ -1,12 +1,10 @@
-local chad_modules = {
-   "options",
-   "mappings",
+local init_modules = {
+   "core",
 }
 
-for i = 1, #chad_modules, 1 do
-   if not pcall(require, chad_modules[i]) then
-      error("Error loading " .. chad_modules[i] .. "\n")
+for _, module in ipairs(init_modules) do
+   local ok, err = pcall(require, module)
+   if not ok then
+      error("Error loading " .. module .. "\n\n" .. err)
    end
 end
-
-require("mappings").misc()
