+++ lua/colors/themes/onedark.lua
