+++ lua/core/options.lua
@@ -1,0 +1,74 @@
+local opt = vim.opt
+local g = vim.g
+
+-- export user config as a global varibale
+g.nvchad_user_config = "chadrc"
+
+local options = require("core.utils").load_config().options
+
+opt.clipboard = options.clipboard
+opt.cmdheight = options.cmdheight
+opt.completeopt = { "menuone", "noselect" }
+opt.cul = true -- cursor line
+
+-- Indentline
+opt.expandtab = options.expandtab
+opt.shiftwidth = options.shiftwidth
+opt.smartindent = options.smartindent
+
+-- disable tilde on end of buffer: https://github.com/  neovim/neovim/pull/8546#issuecomment-643643758
+opt.fillchars = { eob = " " }
+
+opt.hidden = options.hidden
+opt.ignorecase = options.ignorecase
+opt.mouse = options.mouse
+
+-- Numbers
+opt.number = options.number
+opt.numberwidth = options.numberwidth
+opt.relativenumber = options.relativenumber
+opt.ruler = options.ruler
+
+-- disable nvim intro
+opt.shortmess:append "sI"
+opt.signcolumn = "yes"
+opt.splitbelow = true
+opt.splitright = true
+opt.termguicolors = true
+opt.timeoutlen = options.timeoutlen
+opt.undofile = options.permanent_undo
+
+-- interval for writing swap file to disk, also used by gitsigns
+opt.updatetime = options.updatetime
+
+-- go to previous/next line with h,l,left arrow and right arrow
+-- when cursor reaches end/beginning of line
+opt.whichwrap:append "<>hl"
+
+g.mapleader = options.mapleader
+
+-- disable some builtin vim plugins
+local disabled_built_ins = {
+   "2html_plugin",
+   "getscript",
+   "getscriptPlugin",
+   "gzip",
+   "logipat",
+   "netrw",
+   "netrwPlugin",
+   "netrwSettings",
+   "netrwFileHandlers",
+   "matchit",
+   "tar",
+   "tarPlugin",
+   "rrhelper",
+   "spellfile_plugin",
+   "vimball",
+   "vimballPlugin",
+   "zip",
+   "zipPlugin",
+}
+
+for _, plugin in pairs(disabled_built_ins) do
+   g["loaded_" .. plugin] = 1
+end
