+++ lua/plugins/configs/autopairs.lua
@@ -7,6 +7,6 @@ end
 
 autopairs.setup()
 autopairs_completion.setup {
-   map_cr = true,
    map_complete = true, -- insert () func completion
+   map_cr = true,
 }
