+++ lua/plugins/configs/dashboard.lua
@@ -1,7 +1,7 @@
 local g = vim.g
 local fn = vim.fn
 
-local plugins_count = fn.len(fn.globpath("~/.local/share/nvim/site/pack/packer/start", "*", 0, 1))
+-- local plugins_count = fn.len(fn.globpath("~/.local/share/nvim/site/pack/packer/start", "*", 0, 1))
 
 g.dashboard_disable_at_vimenter = 1 -- dashboard is disabled by default
 g.dashboard_disable_statusline = 1
