+++ lua/plugins/configs/gitsigns.lua
@@ -4,18 +4,10 @@ if not present then
 end
 
 gitsigns.setup {
-   signs = {
-      add = { hl = "DiffAdd", text = "│", numhl = "GitSignsAddNr" },
-      change = { hl = "DiffChange", text = "│", numhl = "GitSignsChangeNr" },
-      delete = { hl = "DiffDelete", text = "_", numhl = "GitSignsDeleteNr" },
-      topdelete = { hl = "DiffDelete", text = "‾", numhl = "GitSignsDeleteNr" },
-      changedelete = { hl = "DiffChange", text = "~", numhl = "GitSignsChangeNr" },
-   },
-   numhl = false,
    keymaps = {
       -- Default keymap options
-      noremap = true,
       buffer = true,
+      noremap = true,
       ["n ]c"] = { expr = true, "&diff ? ']c' : '<cmd>lua require\"gitsigns\".next_hunk()<CR>'" },
       ["n [c"] = { expr = true, "&diff ? '[c' : '<cmd>lua require\"gitsigns\".prev_hunk()<CR>'" },
       ["n <leader>hs"] = '<cmd>lua require"gitsigns".stage_hunk()<CR>',
@@ -24,9 +16,19 @@ gitsigns.setup {
       ["n <leader>hp"] = '<cmd>lua require"gitsigns".preview_hunk()<CR>',
       ["n <leader>hb"] = '<cmd>lua require"gitsigns".blame_line()<CR>',
    },
+   numhl = false,
+
+   sign_priority = 5,
+   signs = {
+      add = { hl = "DiffAdd", text = "│", numhl = "GitSignsAddNr" },
+      change = { hl = "DiffChange", text = "│", numhl = "GitSignsChangeNr" },
+      changedelete = { hl = "DiffChange", text = "~", numhl = "GitSignsChangeNr" },
+      delete = { hl = "DiffDelete", text = "_", numhl = "GitSignsDeleteNr" },
+      topdelete = { hl = "DiffDelete", text = "‾", numhl = "GitSignsDeleteNr" },
+   },
+
+   status_formatter = nil, -- Use default
    watch_index = {
       interval = 100,
    },
-   sign_priority = 5,
-   status_formatter = nil, -- Use default
 }
