+++ lua/plugins/configs/telescope.lua
@@ -65,21 +65,23 @@ telescope.setup {
    },
 }
 
--- NvChad pickers
--- load the theme_switcher extension
-require("telescope").load_extension "themes"
--- load the term_picker extension
-require("telescope").load_extension "terms"
+if
+   not pcall(function()
+      -- NvChad pickers
+      -- load the theme_switcher extension
+      telescope.load_extension "themes"
+      -- load the term_picker extension
+      telescope.load_extension "terms"
 
-if not pcall(function()
-   telescope.load_extension "fzf"
-   telescope.load_extension "media_files"
-end) then
+      telescope.load_extension "fzf"
+      telescope.load_extension "media_files"
+   end)
+then
    -- This should only trigger when in need of PackerSync, so better do it
    print "After completion of PackerCompile, restart neovim."
    -- Trigger packer compile on PackerComplete, so it properly waits for PackerSync
    vim.cmd 'autocmd User PackerComplete ++once lua print "Waiting for PackerCompile.." require("packer").compile()'
    vim.cmd 'autocmd User PackerCompileDone ++once echo "Packer Compile done, restart neovim."'
-   require "pluginList"
-   require("packer").update("telescope-fzf-native.nvim", "telescope-media-files.nvim")
+   require "plugins"
+   require("packer").update("core", "telescope-fzf-native.nvim", "telescope-media-files.nvim")
 end
